+++ src/App.jsx
@@ -6,7 +6,7 @@ import Team from "../src/Pages/AboutPage/Team";
 import WhatSpootChat from "../src/Pages/AboutPage/WhatSpootChat";
 import Contact from "../src/Pages/ContactPage/Contact";
 import AccessTo from "./Pages/AccessPage/AccessForm";
-import UserPage from "./Pages/UserPage/UserPage";
+import UserProfile from "./Pages/UserPage/UserProfile";
 import PoliciesAndTerms from "./Pages/MeetPage/PoliciesAndTerms";
 import Support from "./Pages/SupportPage/Support"
 
@@ -21,7 +21,7 @@ export default function App() {
       <Route path="/privacy-policy-and-terms-of-use" element={<PoliciesAndTerms />} />
       <Route path="/contact-us" element={<Contact />} />
       <Route path="/access-to" element={<AccessTo />} />
-      <Route path="/user" element={<UserPage />} />
+      <Route path="/user" element={<UserProfile />} />
       <Route path="/support" element={<Support />} />
       {/* <Route path="/manage-my-account" element={<PonerPaginaCorrespondiente />} /> */}
     </Routes>
