+++ src/Pages/UserPage/UserProfile.jsx
@@ -1,8 +1,8 @@
 import BaseLayout from "../../Components/BaseLayout";
 import { songs } from "../../data";
-import Card from "./Card";
+import Card from "./components/Card";
 
-export default function UserPage (){
+export default function UserProfile (){
 
     return(
         <BaseLayout>
