+++ src/Pages/UserPage/components/Card.jsx
@@ -1,8 +1,6 @@
 import { BsFillPlayCircleFill } from 'react-icons/bs';
 
 export default function Card({ img, name, info, artist }) {
-
-
   return (
     <div>
       <div className="relative flex w-46 flex-col rounded-xl bg-neutral-900 bg-clip-border  shadow-md hover:bg-neutral-900 dark:bg-stone-900 dark:border-neutral-800 dark:hover:bg-neutral-800">
@@ -15,7 +13,6 @@ export default function Card({ img, name, info, artist }) {
             <path strokeLinecap="round" strokeLinejoin="round" d="M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
             <path strokeLinecap="round" strokeLinejoin="round" d="M15.91 11.672a.375.375 0 010 .656l-5.603 3.113a.375.375 0 01-.557-.328V8.887c0-.286.307-.466.557-.327l5.603 3.112z" />
           </svg>
-
         </div>
         <div className="p-6 text-center">
           <h4 className="mb-2 block font-sans text-base font-semibold leading-snug tracking-normal text-white antialiased">
